Fix URLs to remove /poc where not wanted and move /poc stuff out of OpportunityController
