+++ app/views/poc/opportunities/_list.html.haml
@@ -1,7 +1,7 @@
 %dl.opportunities-list
   - list.found.each do |opportunity|
     %dt{:class=>"#{opportunity.source}"}
-      = link_to list.title_with_country(opportunity), poc_opportunity_path(opportunity)
+      = link_to list.title_with_country(opportunity), opportunity_path(opportunity)
 
     %dd{class: "#{opportunity.source}"}
       %span.text
