+++ app/views/poc/opportunities/create.html.haml
@@ -1,0 +1,5 @@
+.container
+  %h1 TODO: POC opportunity#create
+  %p
+    You wish to edit... 
+    =params[:id]
