+++ app/views/poc/opportunities/edit.html.haml
@@ -1,0 +1,5 @@
+.container
+  %h1 TODO: POC opportunity#edit
+  %p
+    You wish to edit... 
+    =params[:id]
