+++ app/views/poc/opportunities/new/_step_1.html.haml
@@ -1,4 +1,4 @@
-%form{:method=>'post'}
+%form{:method=>'post', :action=>poc_opportunities_path}
   %h1= form.title
   %p.description.intro= form.description
