+++ app/views/poc/opportunities/new/_step_2.html.haml
@@ -1,4 +1,4 @@
-%form{:method=>'post'}
+%form{:method=>'post', :action=>poc_opportunities_path}
   = form.hidden_fields
 
   %h1
